Reorganisation of applicationversion service
