+++ src/Marvin.Web/Areas/_ViewImports.cshtml
@@ -1,6 +1,7 @@
 ﻿@using Microsoft.AspNetCore.Identity
 @using Marvin.Web
 @using Marvin.Web.Code.Bootstrap
+@using Marvin.Web.Code.Application
 @using Marvin.Web.Data
 @namespace Marvin.Web.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
